README.md modified

+++ src/api/images/images.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, Options } from '@nestjs/common';
 import { Equal, Raw, Repository } from 'typeorm';
 import * as fs from 'fs';
 
@@ -9,35 +9,36 @@ import { UploadImageDto } from '../dtos/upload-image.dto';
 import { UserImageDto } from '../dtos/user-image.dto';
 import { TotalImageDto } from '../dtos/total-image.dto';
 import { CommentDto } from '../dtos/comment.dto';
+import { UpdateImageDto } from '../dtos/update-image.dto';
 
 @Injectable()
 export class ImagesService {
   // constructor(@InjectRepository(Image) private repo: Repository<Image>) {}
   constructor(@Inject('IMAGE_REPOSITORY') private imageRepository: Repository<Image>, @Inject('COMMENT_REPOSITORY') private commentRepository: Repository<Comment>) {}
 
-  async upload(file: Express.Multer.File, uploadImageData: UploadImageDto): Promise<Result> {
-    const { userId, type } = uploadImageData;
+  async upload(file: Express.Multer.File, uploadImageData: UploadImageDto, req: any): Promise<Result> {
+    const { type } = uploadImageData;
 
     try {
       const newImage = await this.imageRepository.save({
-        userId,
+        userIdx: req.idx,
         filename: file.filename,
         type,
       });
       return { status: 0, msg: '이미지가 성공적으로 업로드 되었습니다.', data: newImage };
     } catch (e) {
-      return { status: 3, msg: `이미지 업로드간 예기치 않은 오류가 발생하였습니다. [${e}]` };
+      return { status: 1, msg: `이미지 업로드간 예기치 않은 오류가 발생하였습니다. [${e}]` };
     }
   }
 
-  async uploads(files: Array<Express.Multer.File>, uploadImageData: UploadImageDto): Promise<Result> {
-    const { userId, type } = uploadImageData;
+  async uploads(files: Array<Express.Multer.File>, uploadImageData: UploadImageDto, req: any): Promise<Result> {
+    const { type } = uploadImageData;
 
     if (files['images'].length > 0) {
       if (Array.isArray(files['images'])) {
         const images = files['images'].map((e) => {
           return {
-            userId,
+            userIdx: req.idx,
             filename: e.filename,
             type,
           };
@@ -53,7 +54,7 @@ export class ImagesService {
       } else {
         try {
           const newImage = await this.imageRepository.save({
-            userId,
+            userIdx: req.idx,
             filename: files['images'][0].filename,
             type,
           });
@@ -71,8 +72,8 @@ export class ImagesService {
     }
   }
 
-  async userImage(userId: string, userImageData: UserImageDto): Promise<Result> {
-    const { date, type, sort } = userImageData;
+  async userImage(userImageData: UserImageDto): Promise<Result> {
+    const { userIdx, date, type, sort } = userImageData;
 
     let images: Array<any>;
     try {
@@ -81,10 +82,10 @@ export class ImagesService {
           where: {
             type: Equal(type),
             createdAt: Raw((alias) => `${alias} > :startDate and ${alias} <= datetime(:endDate, '+1 days')`, { startDate: date, endDate: date }), // sqlite
-            userId,
+            userIdx: Equal(Number(userIdx)),
           },
           order: {
-            id: 'DESC',
+            idx: 'DESC',
           },
         });
       } else {
@@ -92,10 +93,11 @@ export class ImagesService {
           where: {
             type: Equal(type),
             createdAt: Raw((alias) => `${alias} > :startDate and ${alias} <= datetime(:endDate, '+1 days')`, { startDate: date, endDate: date }), // sqlite
-            userId,
+            userIdx: Equal(Number(userIdx)),
           },
         });
       }
+
       return { status: 0, msg: '업로드한 이미지 목록을 출력합니다.', data: images };
     } catch (e) {
       return { status: 1, msg: `업로드한 이미지가 없습니다. [${e}]` };
@@ -119,7 +121,7 @@ export class ImagesService {
             createdAt: Raw((alias) => `${alias} > :startDate and ${alias} <= datetime(:endDate, '+1 days')`, { startDate: date, endDate: date }), // sqlite
           },
           order: {
-            id: 'DESC',
+            idx: 'DESC',
           },
           skip: (Number(num) - 1) * Number(offset),
           take: Number(offset),
@@ -139,41 +141,51 @@ export class ImagesService {
     }
   }
 
-  async updateImage(file: Express.Multer.File, imageId: number, type: string): Promise<Result> {
+  async updateImage(file: Express.Multer.File, imageIdx: number, updateImageData: UpdateImageDto, req: any): Promise<Result> {
+    const { type } = updateImageData;
     const image = await this.imageRepository.findOne({
-      where: { id: imageId },
+      where: { idx: imageIdx },
     });
     if (!image) return { status: 1, msg: '해당 이미지가 없습니다.' };
 
     try {
-      await this.imageRepository.update(image.id, {
-        filename: file.filename,
-        type: type['type'],
-        updatedAt: new Date(),
-      });
+      if (image.userIdx === req.idx) {
+        let option = {
+          updatedAt: new Date(),
+        };
+        if (file) option['filename'] = file.filename;
+        if (type) option['type'] = type;
 
-      const newImage = await this.imageRepository.findOne({
-        where: { id: imageId },
-      });
+        await this.imageRepository.update(image.idx, option);
 
-      fs.unlinkSync(`./uploads/${image.filename}`);
-      return { status: 0, msg: '이미지 정보가 업데이트 되었습니다.', data: newImage };
+        try {
+          if (file) fs.unlinkSync(`./uploads/${image.filename}`);
+        } catch (e) {
+          return { status: 4, msg: `이미지 정보 업데이트간 예기치 않은 오류가 발생하였습니다. [${e}]` };
+        }
+
+        const newImage = await this.imageRepository.findOne({
+          where: { idx: imageIdx },
+        });
+
+        return { status: 0, msg: '이미지 정보가 업데이트 되었습니다.', data: newImage };
+      } else {
+        return { status: 2, msg: '이미지 수정은 본인만 가능합니다.' };
+      }
     } catch (e) {
-      return { status: 1, msg: `이미지 정보 업데이트간 예기치 않은 오류가 발생하였습니다. [${e}]` };
+      return { status: 3, msg: `이미지 정보 업데이트간 예기치 않은 오류가 발생하였습니다. [${e}]` };
     }
   }
-
-  async deleteImage(imageId: number, userId: string): Promise<Result> {
+  async deleteImage(imageIdx: number, req: any): Promise<Result> {
     try {
       const image = await this.imageRepository.findOne({
-        where: { id: imageId },
+        where: { idx: imageIdx },
       });
 
-      // 보안 처리 추가해야 함
-      if (image.userId === userId['userId']) {
+      if (image.userIdx === req.idx) {
         fs.unlinkSync(`./uploads/${image.filename}`);
 
-        await this.imageRepository.delete({ id: imageId });
+        await this.imageRepository.delete({ idx: imageIdx });
         return { status: 0, msg: '이미지가 정상적으로 삭제되었습니다.' };
       } else {
         return { status: 1, msg: '이미지 삭제는 본인만 가능합니다.' };
@@ -183,13 +195,13 @@ export class ImagesService {
     }
   }
 
-  async addComment(imageId: number, commentData: CommentDto): Promise<Result> {
-    const { userId, comment } = commentData;
+  async addComment(imageIdx: number, commentData: CommentDto, req: any): Promise<Result> {
+    const { comment } = commentData;
 
     try {
       const newComment = await this.commentRepository.save({
-        imageId,
-        userId,
+        imageIdx,
+        userIdx: req.idx,
         comment,
       });
       return { status: 0, msg: '이미지에 댓글이 입력되었습니다.', data: newComment };
@@ -198,14 +210,14 @@ export class ImagesService {
     }
   }
 
-  async comment(imageId: number): Promise<Result> {
+  async comment(imageIdx: number): Promise<Result> {
     try {
       const comment = await this.commentRepository.find({
         where: {
-          imageId,
+          imageIdx,
         },
         order: {
-          id: 'DESC',
+          idx: 'DESC',
         },
       });
       return { status: 0, msg: '현재 이미지의 댓글을 출력합니다.', data: comment };
@@ -214,45 +226,41 @@ export class ImagesService {
     }
   }
 
-  async modifyComment(imageId: number, commentId: number, commentData: CommentDto): Promise<Result> {
-    const { userId, comment } = commentData;
+  async modifyComment(imageIdx: number, commentIdx: number, commentData: CommentDto, req: any): Promise<Result> {
+    const { comment } = commentData;
     const modifyComment = await this.commentRepository.findOne({
-      where: { id: commentId, imageId },
+      where: { idx: commentIdx, imageIdx },
     });
+
     if (!modifyComment) return { status: 1, msg: '해당 댓글이 없습니다.' };
     try {
-      // 보안 처리 추가해야 함
-      if (modifyComment.userId === userId) {
-        await this.commentRepository.update(modifyComment.id, {
-          comment: comment['comment'],
+      if (modifyComment.userIdx === req.idx) {
+        await this.commentRepository.update(modifyComment.idx, {
+          comment,
           updatedAt: new Date(),
         });
-  
+
         const newComment = await this.commentRepository.findOne({
-          where: { id: commentId, imageId },
+          where: { idx: commentIdx, imageIdx },
         });
-  
+
         return { status: 0, msg: '댓글 정보가 업데이트 되었습니다.', data: newComment };
       } else {
         return { status: 2, msg: '댓글 수정은 본인만 가능합니다.' };
       }
-
-      
     } catch (e) {
       return { status: 3, msg: `댓글 수정간 예기치 않은 오류가 발생하였습니다. [${e}]` };
     }
   }
 
-  async removeComment(imageId: number, commentId: number, commentData: CommentDto): Promise<Result> {
-    const { userId } = commentData;
+  async removeComment(imageIdx: number, commentIdx: number, req: any): Promise<Result> {
     try {
       const removeComment = await this.commentRepository.findOne({
-        where: { id: commentId, imageId },
+        where: { idx: commentIdx, imageIdx },
       });
 
-      // 보안 처리 추가해야 함
-      if (removeComment.userId === userId) {
-        await this.commentRepository.delete({ id: commentId });
+      if (removeComment.userIdx === req.idx) {
+        await this.commentRepository.delete({ idx: commentIdx });
         return { status: 0, msg: '댓글이 정상적으로 삭제되었습니다.' };
       } else {
         return { status: 1, msg: '댓글 삭제는 본인만 가능합니다.' };
